requirements.txt
discord.py
google-adk
python-dotenv
